requirements/local.txt
-r base.txt
black==24.10.0
isort==5.13.2
pre-commit==4.0.1
django-debug-toolbar==6.0.0
factory_boy==2.12.0
pytest==5.3.2
pytest-django==3.7.0
